Move window ref hashing helpers to Window header
